Refactor: Relocate model classes to `dev.aurakai.auraframefx.models` package
This commit addresses numerous compilation errors by standardizing the location of data model classes. All model-related files, such as `AgentType`, `AgentResponse`, and `AiRequest`, have been moved from the `dev.aurakai.auraframefx.model` package to the new `dev.aurakai.auraframefx.models` package.

Key changes:
- Moved dozens of model and state classes to the `dev.aurakai.auraframefx.models` package.
- Updated all import statements across the application and test files to reflect the new package structure.
- Renamed `AiRequest.kt` to `AiRequestAnnotation.kt` to resolve a class name conflict.
- Downgraded Android Gradle Plugin from `9.0.0-beta02` to `9.0.0-beta01` and upgraded Gradle wrapper to `9.4.0-milestone-2`.
- Removed `AiRequest.kt` (which only contained an annotation) as it was made redundant by the rename.
- Performed minor clean-up in test files, removing unused imports and trailing spaces.
