+++ app/src/main/java/dev/aurakai/auraframefx/ai/agents/AuraShieldAgent.kt
@@ -2,12 +2,11 @@ package dev.aurakai.auraframefx.ai.agents
 
 import android.content.Context
 import dev.aurakai.auraframefx.ai.context.ContextManager
-import dev.aurakai.auraframefx.model.AgentResponse
+import dev.aurakai.auraframefx.models.AgentResponse
 import dev.aurakai.auraframefx.models.AiRequest
 import dev.aurakai.auraframefx.models.agent_states.ActiveThreat
 import dev.aurakai.auraframefx.models.agent_states.ScanEvent
 import dev.aurakai.auraframefx.models.agent_states.SecurityContextState
-import dev.aurakai.auraframefx.models.AiRequest
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
